internal/grpc: remove xDS wrapper types
Updates #1091

Prior to this PR a wrapper type in internal/grpc added the TypeURL
method and sorted the resource values before passing them back up to the
stream handler.

This PR completes pushing TypeURL and Values sorting down to the
grpc.Resource implementations in internal/contour.
